@@ -7,7 +7,7 @@
 		"ethereum",
 		"litecoin"
 	],
-	"version": "1.2.0",
+	"version": "1.3.0",
 	"publisher": "calvert",
 	"repository": {
 		"type": "git",
@@ -86,10 +86,12 @@
 	"devDependencies": {
 		"@types/node": "^10.12.21",
 		"@types/vscode": "^1.37.0",
-		"typescript": "^3.3.1",
-		"tslint": "^5.12.1"
+		"@typescript-eslint/eslint-plugin": "^5.12.0",
+		"@typescript-eslint/parser": "^5.12.0",
+		"typescript": "^4.5.5",
+		"eslint": "^8.9.0"
 	},
 	"dependencies": {
-		"got": "^9.6.0"
+		"got": "^11.8.3"
 	}
 }
